[arc] Move isARCInertBB into ProgramTerminationAnalysis.cpp from ARCAnalysis.cpp.
